+++ lib/client_server_sockets.dart
@@ -1,10 +1,5 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library client_server_sockets;
 
 export 'src/client.dart';
 export 'src/server.dart';
 export 'src/payload.dart';
-
-// TODO: Export any libraries intended for clients of this package.